Refactor end to end test
