+++ tests/end_to_end_tests/test_cli.py
@@ -1,9 +1,11 @@
-import dataregistry_cli.cli as cli
-from dataregistry.db_basic import SCHEMA_VERSION
-from dataregistry import DataRegistry
 import shlex
-from test_end_to_end_python_api import dummy_file
+
+import dataregistry_cli.cli as cli
 import pytest
+from dataregistry import DataRegistry
+from dataregistry.db_basic import SCHEMA_VERSION
+
+from database_test_utils import dummy_file
 
 
 def test_simple_query(dummy_file):
